Create Luna III
my laptop randomly restarted and I was so terrified that I lost everything in this file, so I'm playing it safe and uploading only part of it now. It's 3 am, I will go to sleep and finish it when I wake up like I usually do. I certainly need to learn more about how github works and what tools I need to use.